Refactor repo to reflect the rename.
The original wiki-changes utility was written as a standalone
utility. After writing it, I found myself wanting to write a
program that could benefit from the core deserialization logic
in wiki-changes.

Rather than pulling that code into its own repo, I decided to
repurpose the repo to house all my fedwiki related programs that
I write in Rust.

This checkin preps the codebase for the addition of a second
utility.

It also fixes a bug. --pod no longer requires a parameter.