Initial vesion of wiki-print utility.
wiki-print is intended to be used to export a fedwiki site into a
set of web resources suitable for printing.

This first version can write a bare bones html file for the first
page in the sitemap and open a web browser on Windows to display
it.

Lots more work is neede before it is of much use to anyone.